fix(xm-ace-editor): Fix object to JSON support + formating After YAML support was added, JSON support broke. - Add fixes to support object to JSON convertation - Rollback JSON formating - Replace strings to Enum - Move Interface and Enum to the model file
